CANDataHandler/requirements.txt
asammdf[gui]
watchdog
pyyaml
matplotlib
spicy